+++ lib/appmap/rails/request_handler.rb
@@ -6,15 +6,34 @@ require 'appmap/hook'
 module AppMap
   module Rails
     module RequestHandler
+      IGNORE_HEADERS = Set.new(%w[CONTENT_LENGTH HTTPS HTTP_COOKIE HTTP_HOST HTTP_USER_AGENT PATH_INFO QUERY_STRING REMOTE_ADDR REQUEST_METHOD SCRIPT_NAME SERVER_NAME SERVER_PORT]).freeze
+
+      class << self
+        def selected_headers(headers)
+          matching_headers = headers
+            .to_h
+            .keys
+            .select{ |h| !IGNORE_HEADERS.member?(h) }
+            .select{ |h| !h.index('.') } # Toss out the Rack headers
+          if matching_headers.empty?
+            nil
+          else
+            matching_headers.inject({}) { |memo,key| memo[key] = AppMap::Event::MethodEvent.display_string(headers[key]); memo }
+          end
+        end
+      end
+
       class HTTPServerRequest < AppMap::Event::MethodEvent
-        attr_accessor :normalized_path_info, :request_method, :path_info, :params
+        attr_accessor :normalized_path_info, :request_method, :path_info, :params, :mime_type, :headers
 
         def initialize(request)
           super AppMap::Event.next_id_counter, :call, Thread.current.object_id
 
-          @request_method = request.request_method
-          @normalized_path_info = normalized_path request
-          @path_info = request.path_info.split('?')[0]
+          self.request_method = request.request_method
+          self.normalized_path_info = normalized_path request
+          self.mime_type = request.headers['Content-Type']
+          self.headers = RequestHandler.selected_headers(request.headers)
+          self.path_info = request.path_info.split('?')[0]
           # ActionDispatch::Http::ParameterFilter is deprecated
           parameter_filter_cls = \
             if defined?(ActiveSupport::ParameterFilter)
@@ -22,7 +41,7 @@ module AppMap
             else
               ActionDispatch::Http::ParameterFilter
             end
-          @params = parameter_filter_cls.new(::Rails.application.config.filter_parameters).filter(request.params)
+          self.params = parameter_filter_cls.new(::Rails.application.config.filter_parameters).filter(request.params)
         end
 
         def to_h
@@ -30,7 +49,9 @@ module AppMap
             h[:http_server_request] = {
               request_method: request_method,
               path_info: path_info,
-              normalized_path_info: normalized_path_info
+              mime_type: mime_type,
+              normalized_path_info: normalized_path_info,
+              headers: headers
             }.compact
 
             h[:message] = params.keys.map do |key|
@@ -39,8 +60,11 @@ module AppMap
                 name: key,
                 class: val.class.name,
                 value: self.class.display_string(val),
-                object_id: val.__id__
-              }
+                object_id: val.__id__,
+              }.tap do |message|
+                properties = object_properties(val)
+                message[:properties] = properties if properties
+              end
             end
           end
         end
@@ -59,7 +83,7 @@ module AppMap
       end
 
       class HTTPServerResponse < AppMap::Event::MethodReturnIgnoreValue
-        attr_accessor :status, :mime_type
+        attr_accessor :status, :mime_type, :headers
 
         def initialize(response, parent_id, elapsed)
           super AppMap::Event.next_id_counter, :return, Thread.current.object_id
@@ -68,13 +92,15 @@ module AppMap
           self.mime_type = response.headers['Content-Type']
           self.parent_id = parent_id
           self.elapsed = elapsed
+          self.headers = RequestHandler.selected_headers(response.headers)
         end
 
         def to_h
           super.tap do |h|
             h[:http_server_response] = {
               status: status,
-              mime_type: mime_type
+              mime_type: mime_type,
+              headers: headers
             }.compact
           end
         end